+++ packages/react-self-serializers/src/index.ts
@@ -1,3 +1,4 @@
+export {AppBridge} from './AppBridge';
 export {createSelfSerializer} from './create-self-serializer';
 export {I18n} from './I18n';
 export {GraphQL} from './GraphQLComponent';