+++ packages/react-self-serializers/I18n.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/I18n';