+++ packages/react-self-serializers/CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2019-08-22
+
 ### Fixed
 
 - Published the server entry point ([#898](https://github.com/Shopify/quilt/pull/898))